feat: Exercises 4.15 - 4.23.

+++ exercises/part-4/bloglist-backend/controllers/blogs.js
@@ -1,14 +1,25 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', {
+      username: 1,
+      name: 1,
+    })
 
   response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog
+    .findById(request.params.id)
+    .populate('user', {
+      username: 1,
+      name: 1,
+    })
 
   if (blog) {
     response.json(blog)
@@ -17,18 +28,31 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+
   const blogToSave = new Blog({
-    ...request.body
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes,
+    user: request.user.id,
   })
 
   const savedBlog = await blogToSave.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const blogToUpdate = {
-    ...request.body
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes,
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(
@@ -43,9 +67,26 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blogToRemove = await Blog.findById(request.params.id)
+
+  if (!blogToRemove){
+    response.status(404).send({ error: 'The blog you want to delete does not exist.' })
+  }
+
+  if (blogToRemove.user.toString() === user.id.toString()) {
+    // Remove blog from the blogs collection
+    await Blog.findByIdAndRemove(request.params.id)
+
+    // Remove blog reference from the users collection
+    user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+    await user.save()
+
+    response.status(204).end()
+  } else {
+    response.status(401).send({ error: 'A user can only delete their own blogs.' })
+  }
 })
 
 module.exports = blogsRouter
